Consolidate extractors
